refactor: ⬆️ remove debug comment and seperate command and handeler
