+++ Src/Core/Application/Features/Article/Command/DeleteArticle/DeleteArticleCommand.cs
@@ -1,0 +1,8 @@
+using Application.Common.Interfaces;
+using MediatR;
+
+namespace Application.Features.Article.Command.DeleteArticle;
+
+using CommandType = IRequest<IResponseWrapper<string>>;
+
+public record DeleteArticleCommand(string articleId, string? sub = default) : CommandType;
