+++ Src/Core/Application/Features/Article/Command/LikeArticle/LikeArticleCommand.cs
@@ -1,0 +1,8 @@
+using Application.Common.Interfaces;
+using MediatR;
+
+namespace Application.Features.Article.Command.LikeArticle;
+
+using CommandType = IRequest<IResponseWrapper<string>>;
+
+public record LikeArticleCommand(string articleId, string sub) : CommandType;
