+++ Src/Core/Application/Features/Article/Command/SavaArticle/SaveArticleCommand.cs
@@ -1,0 +1,8 @@
+using Application.Common.Interfaces;
+using MediatR;
+
+namespace Application.Features.Article.Command.SavaArticle;
+
+using CommandType = IRequest<IResponseWrapper<string>>;
+
+public record SaveArticleCommand(string articleId, string sub) : CommandType;
