+++ Src/Core/Application/Features/Article/Command/UnPublish/UnpublishArticleCommand.cs
@@ -1,0 +1,5 @@
+using Application.Common.Mediator;
+
+namespace Application.Features.Article.Command.UnPublish;
+
+public record UnpublishArticleCommand(string id, string sub) : IRequestWraped<string>;
