+++ Src/Core/Application/Features/Article/Command/UnlikeArticle/UnlikeArticleCommandHandeler.cs
@@ -1,15 +1,12 @@
-using Application.Common.Interfaces;
+﻿using Application.Common.Interfaces;
 using Application.Common.models;
 using Application.Common.Repositories;
 using MediatR;
 
-namespace Application.Features.Article.Command;
+namespace Application.Features.Article.Command.UnlikeArticle;
 
 using ResponseType = IResponseWrapper<string>;
-using CommandType = IRequest<IResponseWrapper<string>>;
 using HandlerType = IRequestHandler<UnlikeArticleCommand, IResponseWrapper<string>>;
-
-public record UnlikeArticleCommand(string articleId, string sub) : CommandType;
 
 public class UnlikeArticleCommandHandeler : HandlerType
 {
@@ -27,13 +24,13 @@ public class UnlikeArticleCommandHandeler : HandlerType
 
     private Task<ResponseType> Like(string id, string sub)
     {
-        var res =  _repo.UnLike(id, sub);
+        var res = _repo.UnLike(id, sub);
 
         if (res.IsFaulted && res.Exception is not null)
             return Task.FromResult<ResponseType>(
-                ResponseWrapper.Error<string>(res.Exception ,$"You are unsuccess to lik article : {id}"));
+                ResponseWrapper.Error<string>(res.Exception, $"You are unsuccess to lik article : {id}"));
 
         return Task.FromResult<ResponseType>(
             ResponseWrapper.Ok($"You are success to lik article : {id}"));
-    } 
+    }
 }
