+++ Src/Core/Application/Features/Article/Command/UpdateArticle/UpdateArticleCommand.cs
@@ -1,17 +1,13 @@
 using MediatR;
 using Application.Common.Interfaces;
 using System.Text.Json.Serialization;
-using Application.Common.models;
-using Application.Common.Repositories;
 using Domain.Entites;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.IdentityModel.Tokens;
 
-namespace Application.Features.Article.Command;
+namespace Application.Features.Article.Command.UpdateArticle;
 
-using ResponseType = IResponseWrapper<string>;
 using CommandType = IRequest<IResponseWrapper<string>>;
-using HandlerType = IRequestHandler<UpdateArticleCommand, IResponseWrapper<string>>;
 
 public class UpdateArticleCommand : CommandType
 {
@@ -35,7 +31,7 @@ public class UpdateArticleCommand : CommandType
         {
             Title = Title,
             AuthorId = Sub,
-            ID = ID.IsNullOrEmpty() ? default: new Guid(Base64UrlEncoder.DecodeBytes(ID)),
+            ID = ID.IsNullOrEmpty() ? default : new Guid(Base64UrlEncoder.DecodeBytes(ID)),
             Content = Content,
             Tags = Tags,
             ContentPreviews = Description,
@@ -58,37 +54,4 @@ public class UpdateArticleCommand : CommandType
         }
         return null;
     }
-}
-
-public class UpdateArticleCommandHandler : HandlerType
-{
-    private readonly IPostRepository postRepo;
-
-    public UpdateArticleCommandHandler(IPostRepository postRepo)
-    {
-        this.postRepo = postRepo;
-    }
-
-    public async Task<ResponseType> Handle(UpdateArticleCommand request, CancellationToken cancellationToken)
-    {
-        var v = request.Validate();
-        if (v is not null)
-        {
-            return ResponseWrapper.Error<string>(v, "Data validation fail");
-        }
-
-        var post = request.PostEntity;
-        var a = postRepo.Update( post, cancellationToken);
-
-        if (a.IsCompletedSuccessfully)
-        {
-            return ResponseWrapper.Ok(
-                request.ID,
-                "Update post successfuly.");
-        }
-
-        return ResponseWrapper.Error<string>(a.Exception, "Error occur when save to database");
-
-    }
-
 }
