+++ Src/Core/Application/Features/Article/Query/GetArticle/GetArticleQuery.cs
@@ -1,0 +1,11 @@
+using Application.Common.Interfaces;
+using MediatR;
+
+namespace Application.Features.Article.Query.GetArticle;
+
+using GetArticleQueryType =
+    IRequest<IResponseWrapper<GetArticleQueryResponse>>;
+
+
+public record GetArticleQuery(string ArticleId, string? sub = default) : GetArticleQueryType;
+
