+++ Src/Core/Application/Features/Article/Query/GetArticle/GetArticleQueryHandler.cs
@@ -1,46 +1,17 @@
-using Application.Common.Interfaces;
+﻿using Application.Common.Interfaces;
 using Application.Common.models;
 using Application.Common.Repositories;
 using MediatR;
 using Microsoft.IdentityModel.Tokens;
 
-namespace Application.Features.Article.Query;
+namespace Application.Features.Article.Query.GetArticle;
 
 using GetArticleQueryResponseType =
     IResponseWrapper<GetArticleQueryResponse>;
 
-using GetArticleQueryType =
-    IRequest<IResponseWrapper<GetArticleQueryResponse>>;
-
 using GetArticleQueryHandlerType =
     IRequestHandler<GetArticleQuery, IResponseWrapper<GetArticleQueryResponse>>;
 
-
-public record GetArticleQuery(string ArticleId, string? sub = default) : GetArticleQueryType;
-
-public record GetArticleQueryResponse
-{
-    // identity
-    public virtual string Id { get; set; } = string.Empty;
-    public virtual string AuthorId { get; set; } = string.Empty;
-
-    // contents
-    public virtual string Title { get; set; } = string.Empty;
-    public virtual string CoverImage { get; set; } = string.Empty;
-    public virtual string Description { get; set; } = string.Empty;
-    public virtual string Content { get; set; } = string.Empty;
-
-    // metadata
-    public virtual bool IsPublished { get; set; }
-    public virtual IList<string> Tags { get; set; } = Array.Empty<string>();
-    public virtual DateTime CreatedTime { get; set; }
-    public virtual DateTime LastUpdatedTime { get; set; }
-
-    // state
-    public virtual int LikeCount { get; set; }
-    public virtual bool? IsLiked { get; set; }
-    public virtual bool? IsSaved { get; set; }
-}
 public class GetArticleQueryHandler : GetArticleQueryHandlerType
 {
     private readonly IPostRepository _repo;
@@ -78,7 +49,7 @@ public class GetArticleQueryHandler : GetArticleQueryHandlerType
             IsLiked = article.LikedBy.Contains(request.sub ?? request.ArticleId),
             IsSaved = article.SavedBy.Contains(request.sub ?? request.ArticleId),
             LikeCount = article.LikedCount
-        }) ;
+        });
     }
 }
